CrewAI-SEO-Agency-main/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seo-agency"
version = "0.1.0"
description = "Autonomous SEO Agency using CrewAI framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "SEO Agency Team"}
]

dependencies = [
    "crewai>=0.16.0",
    "crewai[tools]>=0.16.0",
    "openai>=1.6.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.28.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "moz-api-wrapper>=0.3.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true